backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pillow==10.2.0
boto3==1.34.14
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
email-validator==2.1.0
python-dotenv==1.0.0
authlib==1.3.0
itsdangerous==2.1.2
requests==2.31.0